feature: resolve conversations
