+++ src/components/post/Post.module.scss
@@ -5,9 +5,9 @@
   width: 100%;
   margin-bottom: 20px;
   box-shadow: 0px 8px 16px 0px rgba(34, 60, 80, 0.2);
+
   @include medium {
     min-width: 450px;
-    margin-bottom: 20px;
 
     .post {
       padding: 25px;
@@ -32,8 +32,6 @@
   }
 
   @include large {
-    margin-bottom: 20px;
-
     .post {
       padding: 25px;
     }
