+++ src/pages/posts/Posts.module.scss
@@ -31,6 +31,10 @@
         display: flex;
         justify-content: center;
       }
+
+      &__noPostsMessage {
+        margin-top: 100px;
+      }
     }
   }
